Exclude Radio buttons divider from getting checked
Divider element in the raddio buttons have no "value".
If the form element value to match with the radio buttons is
"nil" or "", then it will be matched against the divider elem
(as its value is "nil").
So the "selected item" in the resulting radio button will be the
divider.
This commit resolves it by excluding dividers from the radio button
item selection.