Allow "described_by" to be given as input argument
If we don't have this, we can still pass it as part of the
"attributes: {xxx}" hash. The problem comes when the field also has an
error, generating another "described_by" and causing a 500 error on
render due to the "aria-describedby" attr being declared twice.

To avoid it we allow to provide "described_by" as an argument, and
concatenating the the error_id to it instead of duplicating the
declaration, avoiding the 500 error on render.